requirements.txt
asgiref==3.8.1
certifi==2024.8.30
cloudinary==1.41.0
Django==5.1.3
django-ckeditor==6.7.1
django-js-asset==2.2.0
PyMySQL==1.1.1
six==1.16.0
sqlparse==0.5.2
tzdata==2024.2
urllib3==2.2.3
